initial push of lil analytics v1.0
old repository had a bad commit management style and some bad decisions. decided to move to a new repository and start clean.
bot is already in a production version, it does what it has to do. now comes the time to write tests and improve a lot of functions, since there are quite a few bottlenecks and bad practices.

+ bot is working as intended
+ toolbox.sh to execute common routines of dev process
+ moved to definitions.py from settings.json in this version